requirements.txt
requests
beautifulsoup4
playwright
pytest
pandas
scikit-learn
matplotlib
seaborn
numpy
fastapi
uvicorn
streamlit
plotly
pillow
transformers
torch
huggingface_hub
sentencepiece
shap
python-multipart
